Add --auto_tag behavior
When --auto_tag is enabled, a semver (semver.org) tag like v1.2.3 expands to:
  :1
  :1.2
  :1.2.3

This allows a workflow by which "${DRONE_TAG}" is passed directly to the tags
paraneter, and this results in tagging the docker images correctly for the
"less specific" versions.  It passes through non-semver tags, to allow passing
e.g. ["$DRONE_TAG", "latest"] for images for which "latest" refers only to the
latest release, even if there are "non-release" pushes in between.